backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
pillow==10.4.0
djangorestframework-simplejwt==5.3.1
requests==2.32.3
faker>=20.0.0
celery==5.4.0
redis==5.0.8
django-celery-results==2.5.1
flower==2.0.1
